patch 8.2.1906: warning for signed/unsigned
Problem:    Warning for signed/unsigned.
Solution:   Use size_t instead of int. (Mike Williams)

+++ src/vim9execute.c
@@ -3530,7 +3530,7 @@ ex_disassemble(exarg_T *eap)
 	    case ISN_CMDMOD:
 		{
 		    char_u  *buf;
-		    int	    len = produce_cmdmods(
+		    size_t  len = produce_cmdmods(
 				  NULL, iptr->isn_arg.cmdmod.cf_cmdmod, FALSE);
 
 		    buf = alloc(len + 1);
